Refactor the time drag value widget code to make it more globally available
