+++ crates/viewer/re_selection_panel/src/visible_time_range_ui.rs
@@ -1,10 +1,8 @@
 use std::collections::HashSet;
-use std::ops::RangeInclusive;
 
-use egui::{NumExt as _, Response, Ui};
+use egui::{NumExt as _, Ui};
 
-use re_entity_db::TimeHistogram;
-use re_log_types::{EntityPath, ResolvedTimeRange, TimeType, TimeZone, TimelineName};
+use re_log_types::{EntityPath, ResolvedTimeRange, TimeType, TimelineName};
 use re_space_view_dataframe::DataframeSpaceView;
 use re_space_view_spatial::{SpatialSpaceView2D, SpatialSpaceView3D};
 use re_space_view_time_series::TimeSeriesSpaceView;
@@ -14,7 +12,7 @@ use re_types::{
     Archetype, SpaceViewClassIdentifier,
 };
 use re_ui::UiExt as _;
-use re_viewer_context::{QueryRange, SpaceViewClass, SpaceViewState, ViewerContext};
+use re_viewer_context::{QueryRange, SpaceViewClass, SpaceViewState, TimeDragValue, ViewerContext};
 use re_viewport_blueprint::{entity_path_for_view_property, SpaceViewBlueprint};
 
 /// These space views support the Visible History feature.
@@ -199,18 +197,18 @@ Notes:
                         "Set query range settings for this entity"
                     });
             });
-            let timeline_spec =
+            let time_drag_value =
                 if let Some(times) = ctx.recording().time_histogram(time_ctrl.timeline()) {
-                    TimelineSpec::from_time_histogram(times)
+                    TimeDragValue::from_time_histogram(times)
                 } else {
-                    TimelineSpec::from_time_range(0..=0)
+                    TimeDragValue::from_time_range(0..=0)
                 };
 
             let current_time = TimeInt(
                 time_ctrl
                     .time_i64()
                     .unwrap_or_default()
-                    .at_least(*timeline_spec.range.start()),
+                    .at_least(*time_drag_value.range.start()),
             ); // accounts for timeless time (TimeInt::MIN)
 
             if *has_individual_time_range {
@@ -233,7 +231,7 @@ Notes:
                     current_time,
                     &mut interacting_with_controls,
                     time_type,
-                    &timeline_spec,
+                    &time_drag_value,
                 );
             } else {
                 match &query_range {
@@ -279,7 +277,7 @@ fn time_range_editor(
     current_time: TimeInt,
     interacting_with_controls: &mut bool,
     time_type: TimeType,
-    timeline_spec: &TimelineSpec,
+    time_drag_value: &TimeDragValue,
 ) {
     let current_start = resolved_range.start.start_boundary_time(current_time);
     let current_end = resolved_range.end.end_boundary_time(current_time);
@@ -294,7 +292,7 @@ fn time_range_editor(
                     &mut resolved_range.start,
                     time_type,
                     current_time,
-                    timeline_spec,
+                    time_drag_value,
                     true,
                     current_end,
                 )
@@ -311,7 +309,7 @@ fn time_range_editor(
                     &mut resolved_range.end,
                     time_type,
                     current_time,
-                    timeline_spec,
+                    time_drag_value,
                     false,
                     current_start,
                 )
@@ -476,7 +474,7 @@ fn visible_history_boundary_ui(
     visible_history_boundary: &mut TimeRangeBoundary,
     time_type: TimeType,
     current_time: TimeInt,
-    timeline_spec: &TimelineSpec,
+    time_drag_value: &TimeDragValue,
     low_bound: bool,
     other_boundary_absolute: TimeInt,
 ) -> bool {
@@ -549,8 +547,8 @@ fn visible_history_boundary_ui(
 
             match time_type {
                 TimeType::Time => Some(
-                    timeline_spec
-                        .temporal_drag_value(
+                    time_drag_value
+                        .temporal_drag_value_ui(
                             ui,
                             value,
                             false,
@@ -564,8 +562,8 @@ fn visible_history_boundary_ui(
                         ),
                 ),
                 TimeType::Sequence => Some(
-                    timeline_spec
-                        .sequence_drag_value(ui, value, false, low_bound_override)
+                    time_drag_value
+                        .sequence_drag_value_ui(ui, value, false, low_bound_override)
                         .on_hover_text(
                             "Number of frames before/after the current time to use a time \
                         range boundary",
@@ -583,7 +581,7 @@ fn visible_history_boundary_ui(
 
             match time_type {
                 TimeType::Time => {
-                    let (drag_resp, base_time_resp) = timeline_spec.temporal_drag_value(
+                    let (drag_resp, base_time_resp) = time_drag_value.temporal_drag_value_ui(
                         ui,
                         value,
                         true,
@@ -598,8 +596,8 @@ fn visible_history_boundary_ui(
                     Some(drag_resp.on_hover_text("Absolute time to use as time range boundary"))
                 }
                 TimeType::Sequence => Some(
-                    timeline_spec
-                        .sequence_drag_value(ui, value, true, low_bound_override)
+                    time_drag_value
+                        .sequence_drag_value_ui(ui, value, true, low_bound_override)
                         .on_hover_text("Absolute frame number to use as time range boundary"),
                 ),
             }
@@ -608,229 +606,4 @@ fn visible_history_boundary_ui(
     };
 
     response.map_or(false, |r| r.dragged() || r.has_focus())
-}
-
-// ---
-
-/// Compute and store various information about a timeline related to how the UI should behave.
-#[derive(Debug)]
-struct TimelineSpec {
-    /// Actual range of logged data on the timelines (excluding timeless data).
-    range: RangeInclusive<i64>,
-
-    /// For timelines with large offsets (e.g. `log_time`), this is a rounded time just before the
-    /// first logged data, which can be used as offset in the UI.
-    base_time: Option<i64>,
-
-    // used only for temporal timelines
-    /// For temporal timelines, this is a nice unit factor to use.
-    unit_factor: i64,
-
-    /// For temporal timelines, this is the unit symbol to display.
-    unit_symbol: &'static str,
-
-    /// This is a nice range of absolute times to use when editing an absolute time. The boundaries
-    /// are extended to the nearest rounded unit to minimize glitches.
-    abs_range: RangeInclusive<i64>,
-
-    /// This is a nice range of relative times to use when editing an absolute time. The boundaries
-    /// are extended to the nearest rounded unit to minimize glitches.
-    rel_range: RangeInclusive<i64>,
-}
-
-impl TimelineSpec {
-    fn from_time_histogram(times: &TimeHistogram) -> Self {
-        Self::from_time_range(
-            times.min_key().unwrap_or_default()..=times.max_key().unwrap_or_default(),
-        )
-    }
-
-    fn from_time_range(range: RangeInclusive<i64>) -> Self {
-        let span = range.end() - range.start();
-        let base_time = time_range_base_time(*range.start(), span);
-        let (unit_symbol, unit_factor) = unit_from_span(span);
-
-        // `abs_range` is used by the DragValue when editing an absolute time, its bound expended to
-        // nearest unit to minimize glitches.
-        let abs_range =
-            round_down(*range.start(), unit_factor)..=round_up(*range.end(), unit_factor);
-
-        // `rel_range` is used by the DragValue when editing a relative time offset. It must have
-        // enough margin either side to accommodate for all possible values of current time.
-        let rel_range = round_down(-span, unit_factor)..=round_up(2 * span, unit_factor);
-
-        Self {
-            range,
-            base_time,
-            unit_factor,
-            unit_symbol,
-            abs_range,
-            rel_range,
-        }
-    }
-
-    fn sequence_drag_value(
-        &self,
-        ui: &mut egui::Ui,
-        value: &mut TimeInt,
-        absolute: bool,
-        low_bound_override: Option<TimeInt>,
-    ) -> Response {
-        let mut time_range = if absolute {
-            self.abs_range.clone()
-        } else {
-            self.rel_range.clone()
-        };
-
-        // speed must be computed before messing with time_range for consistency
-        let span = time_range.end() - time_range.start();
-        let speed = (span as f32 * 0.005).at_least(1.0);
-
-        if let Some(low_bound_override) = low_bound_override {
-            time_range = low_bound_override.0.at_least(*time_range.start())..=*time_range.end();
-        }
-
-        ui.add(
-            egui::DragValue::new(&mut value.0)
-                .range(time_range)
-                .speed(speed),
-        )
-    }
-
-    /// Show a temporal drag value.
-    ///
-    /// Feature rich:
-    /// - scale to the proper units
-    /// - display the base time if any
-    /// - etc.
-    ///
-    /// Returns a tuple of the [`egui::DragValue`]'s [`egui::Response`], and the base time label's
-    /// [`egui::Response`], if any.
-    fn temporal_drag_value(
-        &self,
-        ui: &mut egui::Ui,
-        value: &mut TimeInt,
-        absolute: bool,
-        low_bound_override: Option<TimeInt>,
-        time_zone_for_timestamps: TimeZone,
-    ) -> (Response, Option<Response>) {
-        let mut time_range = if absolute {
-            self.abs_range.clone()
-        } else {
-            self.rel_range.clone()
-        };
-
-        let factor = self.unit_factor as f32;
-        let offset = if absolute {
-            self.base_time.unwrap_or(0)
-        } else {
-            0
-        };
-
-        // speed must be computed before messing with time_range for consistency
-        let speed = (time_range.end() - time_range.start()) as f32 / factor * 0.005;
-
-        if let Some(low_bound_override) = low_bound_override {
-            time_range = low_bound_override.0.at_least(*time_range.start())..=*time_range.end();
-        }
-
-        let mut time_unit = (value.0 - offset) as f32 / factor;
-
-        let time_range = (*time_range.start() - offset) as f32 / factor
-            ..=(*time_range.end() - offset) as f32 / factor;
-
-        let base_time_response = if absolute {
-            self.base_time.map(|base_time| {
-                ui.label(format!(
-                    "{} + ",
-                    TimeType::Time.format(TimeInt(base_time), time_zone_for_timestamps)
-                ))
-            })
-        } else {
-            None
-        };
-
-        let drag_value_response = ui.add(
-            egui::DragValue::new(&mut time_unit)
-                .range(time_range)
-                .speed(speed)
-                .suffix(self.unit_symbol),
-        );
-
-        *value = TimeInt((time_unit * factor).round() as i64 + offset);
-
-        (drag_value_response, base_time_response)
-    }
-}
-
-fn unit_from_span(span: i64) -> (&'static str, i64) {
-    if span / 1_000_000_000 > 0 {
-        ("s", 1_000_000_000)
-    } else if span / 1_000_000 > 0 {
-        ("ms", 1_000_000)
-    } else if span / 1_000 > 0 {
-        ("μs", 1_000)
-    } else {
-        ("ns", 1)
-    }
-}
-
-/// Value of the start time over time span ratio above which an explicit offset is handled.
-static SPAN_TO_START_TIME_OFFSET_THRESHOLD: i64 = 10;
-
-fn time_range_base_time(min_time: i64, span: i64) -> Option<i64> {
-    if min_time <= 0 {
-        return None;
-    }
-
-    if span.saturating_mul(SPAN_TO_START_TIME_OFFSET_THRESHOLD) < min_time {
-        let factor = if span / 1_000_000 > 0 {
-            1_000_000_000
-        } else if span / 1_000 > 0 {
-            1_000_000
-        } else {
-            1_000
-        };
-
-        Some(min_time - (min_time % factor))
-    } else {
-        None
-    }
-}
-
-fn round_down(value: i64, factor: i64) -> i64 {
-    value - (value.rem_euclid(factor))
-}
-
-fn round_up(value: i64, factor: i64) -> i64 {
-    let val = round_down(value, factor);
-
-    if val == value {
-        val
-    } else {
-        val + factor
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_round_down() {
-        assert_eq!(round_down(2200, 1000), 2000);
-        assert_eq!(round_down(2000, 1000), 2000);
-        assert_eq!(round_down(-2200, 1000), -3000);
-        assert_eq!(round_down(-3000, 1000), -3000);
-        assert_eq!(round_down(0, 1000), 0);
-    }
-
-    #[test]
-    fn test_round_up() {
-        assert_eq!(round_up(2200, 1000), 3000);
-        assert_eq!(round_up(2000, 1000), 2000);
-        assert_eq!(round_up(-2200, 1000), -2000);
-        assert_eq!(round_up(-3000, 1000), -3000);
-        assert_eq!(round_up(0, 1000), 0);
-    }
 }
